Rename maps-v3-client to fetch-map
