+++ modules/carto/src/api/index.ts
@@ -1,4 +1,4 @@
 export {CartoAPIError} from './carto-api-error';
-export {fetchMap} from './maps-v3-client';
-export type {FetchMapOptions} from './maps-v3-client';
+export {fetchMap} from './fetch-map';
+export type {FetchMapOptions} from './fetch-map';
 export type {APIErrorContext, Format, MapType, RequestType, QueryParameters} from './types';
